At long last... I've finally done it, sorta
So, many months ago, I tried to make it so that particles would spawn next to the player on both their left and right. I couldn't do it, but about 2 weeks ago, I remembered that the Phantom has particles, which do exactly the thing I wanted to do. So, I've copied the code over, and it works! Sorta. It hasn't been adjusted for the elytra's position, but that can be done later. This mini session here was to prove that it was possible. I hope to come back to this project with fresh ideas, better knowledge, and even some rendering beyond particles and screen overlays, based on Loot Beams. I'm unsure, but I want to improve my rendering skills further, and this project and other idea I've had since the beginning may be the way I begin.

Also I added a new particle, apparently.